refactor: connection string builder and fix incorrect tests (#81)

+++ testframework/src/test/java/host/IntegrationContainerTest.java
@@ -76,13 +76,15 @@ public class IntegrationContainerTest {
   private static final String ENDPOINT = System.getenv("RDS_ENDPOINT");
   private static final String REGION = System.getenv("RDS_REGION");
 
+  private static final boolean REUSE_CLUSTER = Boolean.parseBoolean(System.getenv("REUSE_RDS_CLUSTER"));
+
   private static final String DOCKER_UID = "1001";
   private static final String COMMUNITY_SERVER = "postgres-instance";
 
   private static final String DRIVER_LOCATION = System.getenv("DRIVER_PATH");
   private static final String PROXIED_DOMAIN_NAME_SUFFIX = ".proxied";
   private static List<ToxiproxyContainer> proxyContainers = new ArrayList<>();
-  private static String dbClusterIdentifier = System.getenv("TEST_DB_CLUSTER_IDENTIFIER");
+  private static String dbClusterIdentifier = System.getenv("RDS_CLUSTER_NAME");
   private static String dbShardGroupIdentifier = System.getenv("TEST_DB_SHARD_GROUP_IDENTIFIER");
 
   private static final String ODBCINI_LOCATION = "/app/build/test/odbc.ini";
@@ -116,7 +118,10 @@ public class IntegrationContainerTest {
 
   @AfterAll
   static void tearDown() {
-    if (!StringUtils.isNullOrEmpty(ACCESS_KEY) && !StringUtils.isNullOrEmpty(SECRET_ACCESS_KEY) && !StringUtils.isNullOrEmpty(dbHostCluster)) {
+    if (!REUSE_CLUSTER
+        && !StringUtils.isNullOrEmpty(ACCESS_KEY)
+        && !StringUtils.isNullOrEmpty(SECRET_ACCESS_KEY)
+        && !StringUtils.isNullOrEmpty(dbHostCluster)) {
       switch (testConfiguration) {
         case LIMITLESS:
           auroraUtil.deleteLimitlessCluster(dbClusterIdentifier, dbShardGroupIdentifier);
@@ -382,26 +387,29 @@ public class IntegrationContainerTest {
 
   private void setupApgTestContainer(final Network network) throws InterruptedException, UnknownHostException {
     if (!StringUtils.isNullOrEmpty(ACCESS_KEY) && !StringUtils.isNullOrEmpty(SECRET_ACCESS_KEY)) {
-      // Comment out below to not create a new cluster & instances
+      AuroraClusterInfo clusterInfo;
 
-      if (StringUtils.isNullOrEmpty(dbClusterIdentifier)) {
-        dbClusterIdentifier = DEFAULT_APG_PREFIX + "cluster-" + System.nanoTime();
+      if (REUSE_CLUSTER && !dbClusterIdentifier.isEmpty()) {
+        clusterInfo = auroraUtil.getClusterInfo(dbClusterIdentifier);
+      } else {
+        if (StringUtils.isNullOrEmpty(dbClusterIdentifier)) {
+          dbClusterIdentifier = DEFAULT_APG_PREFIX + "cluster-" + System.nanoTime();
+        }
+
+        clusterInfo = auroraUtil.createApgCluster(TEST_USERNAME, TEST_PASSWORD, dbClusterIdentifier);
+
+        runnerIP = auroraUtil.getPublicIPAddress();
+        auroraUtil.ec2AuthorizeIP(runnerIP);
+
+        String secretValue = auroraUtil.createSecretValue(dbHostCluster, TEST_USERNAME, TEST_PASSWORD);
+        secretsArn = auroraUtil.createSecrets("AWS-PGSQL-ODBC-Tests-" + dbHostCluster, secretValue);
       }
-
-      AuroraClusterInfo clusterInfo =
-          auroraUtil.createApgCluster(TEST_USERNAME, TEST_PASSWORD, dbClusterIdentifier);
-
-      // Comment out getting public IP to not add & remove from EC2 whitelist
-      runnerIP = auroraUtil.getPublicIPAddress();
-      auroraUtil.ec2AuthorizeIP(runnerIP);
 
       dbConnStrSuffix = clusterInfo.getClusterSuffix();
       dbHostCluster = clusterInfo.getClusterEndpoint();
       dbHostClusterRo = clusterInfo.getClusterROEndpoint();
 
       postgresInstances = clusterInfo.getInstances();
-      String secretValue = auroraUtil.createSecretValue(dbHostCluster, TEST_USERNAME, TEST_PASSWORD);
-      secretsArn = auroraUtil.createSecrets("AWS-PGSQL-ODBC-Tests-" + dbHostCluster, secretValue);
 
       proxyContainers = containerHelper.createProxyContainers(network, postgresInstances, PROXIED_DOMAIN_NAME_SUFFIX);
       for (ToxiproxyContainer container : proxyContainers) {
